+++ src/scripts/train/OLMoE3-dev-260401-a0.6s.py
@@ -42,7 +42,7 @@ from olmo_core.nn.transformer import (
 from olmo_core.optim import OptimGroupOverride, SchedulerUnits
 from olmo_core.optim.scheduler import (
     ComposableScheduler,
-    ComposableSchedulerMonkeyPatchDecay,
+    OverrideDecay,
     ComposableSchedulerStage,
     ComposableSchedulerStageType,
 )
@@ -482,8 +482,8 @@ def build_train_module_config(common: CommonComponents) -> MoEV2TransformerTrain
             # duration: 2B tokens
             # start LR: 8.7138e-4
             # end LR:
-            monkey_patch_decay=(
-                ComposableSchedulerMonkeyPatchDecay(
+            override_decay=(
+                OverrideDecay(
                     start=MONKEY_PATCH_DECAY_START_TOKENS,
                     duration=MONKEY_PATCH_DECAY_DURATION_TOKENS,
                     shape=MONKEY_PATCH_DECAY_SHAPE,